pyproject.toml
[project]
name = "document-ai-hub"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pypdf (>=6.2.0,<7.0.0)",
    "sentence-transformers (>=5.1.2,<6.0.0)",
    "chromadb (>=1.3.4,<2.0.0)",
    "ollama (>=0.6.0,<0.7.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "easyocr (>=1.7.2,<2.0.0)",
    "sentencepiece (>=0.2.1,<0.3.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "langchain-ollama (>=1.0.0,<2.0.0)",
    "langchain-core (>=1.0.4,<2.0.0)",
    "streamlit (>=1.51.0,<2.0.0)",
    "streamlit-extras (>=0.7.8,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.1,<10.0.0)",
    "black (>=25.11.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)"
]
